fix: optimize increment merge
