+++ src/service/compact/merge.rs
@@ -18,7 +18,7 @@ use std::sync::Arc;
 use ::datafusion::{arrow::datatypes::Schema, error::DataFusionError};
 use arrow::array::RecordBatch;
 use bytes::Bytes;
-use chrono::{DateTime, Datelike, Duration, TimeZone, Timelike, Utc};
+use chrono::{DateTime, Datelike, Duration, TimeZone, Utc};
 use config::{
     FileFormat, TIMESTAMP_COL_NAME,
     cluster::LOCAL_NODE,
@@ -32,7 +32,7 @@ use config::{
         parquet::{get_recordbatch_reader_from_bytes, read_schema_from_bytes},
         record_batch_ext::concat_batches,
         schema_ext::SchemaExt,
-        time::{day_micros, hour_micros, now_micros},
+        time::{day_micros, hour_micros},
     },
 };
 use hashbrown::{HashMap, HashSet};
@@ -115,34 +115,7 @@ pub async fn generate_job_by_stream(
 
     // format to hour with zero minutes, seconds
     let offset = offset - offset % hour_micros(1);
-
-    let cfg = get_config();
-    // check offset
-    let time_now: DateTime<Utc> = Utc::now();
-    let time_now_hour = Utc
-        .with_ymd_and_hms(
-            time_now.year(),
-            time_now.month(),
-            time_now.day(),
-            time_now.hour(),
-            0,
-            0,
-        )
-        .unwrap()
-        .timestamp_micros();
-    // must wait for at least 3 * max_file_retention_time
-    // -- first period: the last hour local file upload to storage, write file list
-    // -- second period, the last hour file list upload to storage
-    // -- third period, we can do the merge, so, at least 3 times of
-    // max_file_retention_time
-    if offset >= time_now_hour
-        || time_now.timestamp_micros() - offset
-            <= Duration::try_seconds(cfg.limit.max_file_retention_time as i64)
-                .unwrap()
-                .num_microseconds()
-                .unwrap()
-                * 3
-    {
+    if !super::is_past_hour(offset) {
         return Ok(()); // the time is future, just wait
     }
 
@@ -382,29 +355,6 @@ pub async fn generate_downsampling_job_by_stream_and_rule(
     Ok(())
 }
 
-/// Retire a finished merge job. Re-checks at completion whether the job's hour is still
-/// the current (open) hour. If so, it's an incremental round: delete the job row so the
-/// unique (stream, offsets) key is freed for the next round and the dump path never picks
-/// it up. Otherwise the hour has ended: hand off to the normal done (+ dump) path.
-///
-/// The re-check at completion (rather than at start) closes a boundary race: the scheduled
-/// `generate_job_by_stream` advances its per-stream offset cursor unconditionally after
-/// `add_job` (even on an `ON CONFLICT` no-op), so if we deleted a job after the cursor had
-/// already moved past its hour, that hour would never get a done+dump pass. By only
-/// deleting while the hour is still open, a job that straddles the boundary falls through
-/// to `set_job_done` and is dumped normally.
-async fn finish_merge_job(job_id: i64, offset: i64) {
-    let now = now_micros();
-    let is_current_hour = offset >= now - now % hour_micros(1);
-    if is_current_hour {
-        if let Err(e) = infra_file_list::del_jobs(&[job_id]).await {
-            log::error!("[COMPACTOR] del incremental job [{job_id}] failed: {e}");
-        }
-    } else if let Err(e) = infra_file_list::set_job_done(&[job_id]).await {
-        log::error!("[COMPACTOR] set_job_done [{job_id}] failed: {e}");
-    }
-}
-
 /// compactor run steps on a stream:
 /// 3. get a cluster lock for compactor stream
 /// 4. read last compacted offset: year/month/day/hour
@@ -430,7 +380,9 @@ pub async fn merge_by_stream(
     let schema = infra::schema::get(org_id, stream_name, stream_type).await?;
     if schema == Schema::empty() {
         // the stream was deleted, mark the job as done
-        finish_merge_job(job_id, offset).await;
+        if let Err(e) = infra_file_list::set_job_done(&[job_id]).await {
+            log::error!("[COMPACTOR] set_job_done failed: {e}");
+        }
         return Ok(());
     }
 
@@ -442,10 +394,8 @@ pub async fn merge_by_stream(
     // still-open current hour (enqueued by the ingester, see service::compact::incremental):
     // only seal full-size groups and carry the remainder, so each file is merged into a
     // sealed output exactly once. The scheduled hour-end pass seals whatever is left.
-    let is_incremental = {
-        let now = now_micros();
-        offset >= now - now % hour_micros(1)
-    };
+    let offset = offset - offset % hour_micros(1);
+    let is_incremental = !super::is_past_hour(offset);
 
     // check offset
     let partition_time_level = get_partition_time_level(stream_type);
@@ -473,7 +423,9 @@ pub async fn merge_by_stream(
 
     if files.is_empty() {
         // update job status
-        finish_merge_job(job_id, offset).await;
+        if let Err(e) = infra_file_list::set_job_done(&[job_id]).await {
+            log::error!("[COMPACTOR] set_job_done failed: {e}");
+        }
         return Ok(());
     }
 
@@ -661,9 +613,10 @@ pub async fn merge_by_stream(
         task.await??;
     }
 
-    // update job status: incremental (current-hour) jobs are deleted so the next round can
-    // re-trigger and dump skips them; ended hours go through the normal done (+ dump) path
-    finish_merge_job(job_id, offset).await;
+    // update job status
+    if let Err(e) = infra_file_list::set_job_done(&[job_id]).await {
+        log::error!("[COMPACTOR] set_job_done failed: {e}");
+    }
 
     // metrics
     let time = start.elapsed().as_secs_f64();
