+++ src/service/compact/mod.rs
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-use chrono::{Datelike, Duration, TimeZone, Timelike, Utc};
+use chrono::{DateTime, Datelike, Duration, TimeZone, Timelike, Utc};
 use config::{
     COMPACT_OLD_DATA_STREAM_SET,
     cluster::LOCAL_NODE,
@@ -449,4 +449,31 @@ pub async fn run_delay_deletion() -> Result<(), anyhow::Error> {
     }
 
     Ok(())
+}
+
+pub(crate) fn is_past_hour(offset: i64) -> bool {
+    let time_now: DateTime<Utc> = Utc::now();
+    let time_now_hour = Utc
+        .with_ymd_and_hms(
+            time_now.year(),
+            time_now.month(),
+            time_now.day(),
+            time_now.hour(),
+            0,
+            0,
+        )
+        .unwrap()
+        .timestamp_micros();
+    // must wait for at least 3 * max_file_retention_time
+    // -- first period: the last hour local file upload to storage, write file list
+    // -- second period, the last hour file list upload to storage
+    // -- third period, we can do the merge, so, at least 3 times of
+    // max_file_retention_time
+    offset < time_now_hour
+        && time_now.timestamp_micros() - offset
+            > Duration::try_seconds(get_config().limit.max_file_retention_time as i64)
+                .unwrap()
+                .num_microseconds()
+                .unwrap()
+                * 3
 }
